src/app/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.5.0
pandas==2.1.4
numpy==1.26.2
prophet==1.1.5
scikit-learn==1.3.2
python-multipart==0.0.6
python-dotenv==1.0.0
aiofiles==23.2.1
plotly==5.18.0
openpyxl==3.1.5
